pyproject.toml
[project]
name = "chat"
version = "0.5.0"
description = "RAG ChatBot"
authors = [
    { name = "Umberto Griffo", email = "umberto.griffo@gmail.com" }
]
readme = "README.md"

dynamic = ["dependencies", "requires-python"]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
torch = [
    { version = "~=2.5.1", source = "pytorch", platform = "!=darwin" },
    { version = "~=2.5.1", source = "PyPI", platform = "darwin" }
]
sentence-transformers = { version = "~=5.0.0" }
sentencepiece = "~=0.2.1"
tqdm = "~=4.67.0"
requests = "~=2.32.0"
numpy = "~=1.26.4"
streamlit = "~=1.37.0"
unstructured = { version = "~=0.20.2", extras = ["md"] }
chromadb = "~=1.5.0"
transformers = "~=4.57.6"
rich = "~=13.4.2"
pyfiglet = "~=0.7"
clean-text = "~=0.7.1"
Unidecode = "~=1.3.6"
nest_asyncio = "~=1.5.8"
llama-cpp-python = "^0.3.16"
anthropic = ">=0.20.0"
openai = ">=1.0.0"
google-generativeai = ">=0.4.0"
python-dotenv = ">=1.0.0"
audio-recorder-streamlit = ">=0.0.8"
openai-whisper = ">=20231117"
python-docx = ">=1.0.0"
openpyxl = ">=3.1.0"
pypdf = ">=4.0.0"
pillow = ">=10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "~=7.2.1"
pytest-cov = "~=4.0.0"
pytest-mock = "~=3.10.0"
pytest-asyncio = "~=0.23.6"
pre-commit = "~=3.6.0"
ruff = "~=0.6.4"
httpx = "~=0.28.1"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# https://docs.astral.sh/ruff/configuration/
# https://docs.astral.sh/ruff/settings/#lint_isort_known-first-party
[tool.ruff]
exclude = [
    ".venv",
    ".ruff_cache",
    ".pytest_cache",
    "docs"
]
line-length = 120
indent-width = 4

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# "I" is isort, C901 is McCabe complexity
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I",
    "C901"
]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.mccabe]
# Flag errors (`C901`) whenever the complexity level exceeds 12.
max-complexity = 12

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.pytest.ini_options]
pythonpath = [
    "chatbot",
    "tests"
]
